Fix empty trash when is trashed only WP Users list
[MAILPOET-3142]

+++ lib/Models/Segment.php
@@ -341,6 +341,9 @@ class Segment extends Model {
       $ids = array_map(function($segment) {
         return $segment->id;
       }, $segments);
+      if (!$ids) {
+        return;
+      }
       SubscriberSegment::whereIn('segment_id', $ids)
         ->deleteMany();
       Segment::whereIn('id', $ids)->deleteMany();
